# packages/api-server/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Rita API",
		"version": "1.0.0",
		"description": "Rita API Server for the RITA application.\n\n## Authentication\nMost endpoints require authentication via Keycloak JWT token or session cookie.\n\n## Organization Context\nAll data is scoped to the user's active organization (tenant).\n\n## Real-time Updates\nFor real-time updates, connect to the SSE endpoint at `/api/sse/events`.\nSee AsyncAPI spec for event documentation.",
		"contact": {
			"name": "Rita Team"
		}
	},
	"servers": [
		{
			"url": "/",
			"description": "Current server"
		}
	],
...
	"security": [
		{
			"bearerAuth": []
		},
		{
			"cookieAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Keycloak JWT token"
			},
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session",
				"description": "Session cookie authentication"
			}
		},
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string",
						"description": "Error message",
						"example": "Resource not found"
					},
					"code": {
						"type": "string",
						"description": "Error code",
						"example": "NOT_FOUND"
					},
					"details": {
						"description": "Additional error details"
					}
				},
				"required": ["error"]
			},
...
			"IframeValidateResponse": {
				"type": "object",
				"properties": {
					"valid": {
						"type": "boolean",
						"description": "Whether validation succeeded"
					},
					"conversationId": {
						"type": "string",
						"format": "uuid",
						"description": "Conversation ID"
					},
					"webhookConfigLoaded": {
						"type": "boolean",
						"description": "Whether webhook config loaded"
					},
					"webhookTenantId": {
						"type": "string",
						"description": "Tenant ID from webhook config"
					},
					"titleText": {
						"type": "string",
						"description": "Custom title text from Valkey"
					},
					"welcomeText": {
						"type": "string",
						"description": "Custom welcome text from Valkey"
					},
					"placeholderText": {
						"type": "string",
						"description": "Custom placeholder text from Valkey"
					},
					"valkeyPayload": {
						"type": "object",
						"additionalProperties": {},
						"description": "Full Valkey payload (dev tools)"
					},
					"error": {
						"type": "string",
						"description": "Error message if validation failed"
					}
				},
				"required": ["valid"]
			},
			"IframeValidateRequest": {
				"type": "object",
				"properties": {
					"sessionKey": {
						"type": "string",
						"description": "Valkey session key"
					},
					"intentEid": {
						"type": "string",
						"description": "Optional intent EID for tracking"
					},
					"existingConversationId": {
						"type": "string",
						"format": "uuid",
						"description": "Existing conversation ID to resume"
					}
				},
				"required": ["sessionKey"]
			},
...
			"IframeExecuteResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether execution started"
					},
					"eventId": {
						"type": "string",
						"description": "Event ID for tracking"
					},
					"error": {
						"type": "string",
						"description": "Error message if failed"
					},
					"debug": {
						"type": "object",
						"properties": {
							"valkeyStatus": {},
							"totalDurationMs": {
								"type": "number"
							},
							"errorCode": {
								"type": "string"
							},
							"errorDetails": {
								"type": "string"
							}
						},
						"description": "Debug information"
					}
				},
				"required": ["success"]
			},
			"IframeExecuteRequest": {
				"type": "object",
				"properties": {
					"hashkey": {
						"type": "string",
						"description": "Valkey hashkey for workflow"
					},
					"sessionKey": {
						"type": "string",
						"description": "Alternative to hashkey (portal uses this)"
					}
				}
			},
			"IframeDeleteConversationResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether deletion succeeded"
					},
					"error": {
						"type": "string",
						"description": "Error message if failed"
					}
				},
				"required": ["success"]
			},
			"IframeSessionContextResponse": {
				"type": "object",
				"properties": {
					"tenantId": {
						"type": "string",
						"description": "Tenant ID from Valkey session"
					},
					"userGuid": {
						"type": "string",
						"description": "User GUID from Valkey session"
					},
					"context": {
						"type": "object",
						"additionalProperties": {},
						"description": "Session context (runId, activityId, etc.)"
					},
					"accessToken": {
						"type": "string",
						"description": "Access token (redacted)"
					},
					"refreshToken": {
						"type": "string",
						"description": "Refresh token (redacted)"
					},
					"clientKey": {
						"type": "string",
						"description": "Client key (redacted)"
					}
				},
				"additionalProperties": {}
			},
...
		},
...
	},
	"paths": {
...
		"/api/iframe/validate-instantiation": {
			"post": {
				"tags": ["Iframe"],
				"summary": "Validate iframe instantiation",
				"description": "Validate Valkey session key and setup iframe session. No auth required - sessionKey is the auth.",
				"security": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IframeValidateRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Validation successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeValidateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Validation failed - invalid session",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeValidateResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/iframe/execute": {
			"post": {
				"tags": ["Iframe"],
				"summary": "Execute workflow",
				"description": "Execute workflow from Valkey hashkey. No auth required - hashkey provides context.",
				"security": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IframeExecuteRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Execution started",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeExecuteResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input or execution failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeExecuteResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeExecuteResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/iframe/conversation/{conversationId}": {
			"delete": {
				"tags": ["Iframe"],
				"summary": "Delete conversation",
				"description": "Delete iframe conversation and all messages. Used for 'Clear Chat' feature. Requires session cookie and conversation ownership.",
				"parameters": [
					{
						"schema": {
							"type": "string",
							"format": "uuid",
							"description": "Conversation ID to delete"
						},
						"required": true,
						"description": "Conversation ID to delete",
						"name": "conversationId",
						"in": "path"
					}
				],
				"responses": {
					"200": {
						"description": "Conversation deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeDeleteConversationResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing conversationId",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeDeleteConversationResponse"
								}
							}
						}
					},
					"500": {
						"description": "Deletion failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeDeleteConversationResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/iframe/session-context": {
			"get": {
				"tags": ["Iframe"],
				"summary": "Get fresh session context from Valkey",
				"description": "Re-reads Valkey and returns fresh payload with sensitive fields redacted. Used before metadata download to get latest context (runId, activityId).",
				"security": [],
				"parameters": [
					{
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 256,
							"description": "Valkey session key"
						},
						"required": true,
						"description": "Valkey session key",
						"name": "sessionKey",
						"in": "query"
					}
				],
				"responses": {
					"200": {
						"description": "Fresh session context",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IframeSessionContextResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing or invalid sessionKey",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Session not found in Valkey",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}